+++ common/config/config.go
@@ -66,7 +66,6 @@ var StripeEndpointSecret = ""
 var StripeApiSecret = ""
 var StripeWebhookSecret = ""
 var StripePriceId = ""
-var StripeUnitPrice = 8.0
 var StripeMinTopUp = 1
 var StripePromotionCodesEnabled = false