+++ router/api-router.go
@@ -56,7 +56,6 @@ func SetApiRouter(router *gin.Engine) {
 				selfRoute.POST("/topup", controller.TopUp)
 				selfRoute.GET("/topup/self", controller.GetUserTopUps)
 				selfRoute.POST("/stripe/pay", middleware.CriticalRateLimit(), controller.RequestStripePay)
-				selfRoute.POST("/stripe/amount", controller.RequestStripeAmount)
 			}
 
 			adminRoute := userRoute.Group("/")